+++ src/components/RequestNewForm.tsx
@@ -1,27 +1,6 @@
 import React, { useEffect, useState } from "react";
-import { useForm } from "react-hook-form";
-import { zodResolver } from "@hookform/resolvers/zod";
-import { z } from "zod";
-import { cn } from "@/lib/utils";
-import { format } from "date-fns";
 import { MessageColors, setMessage } from "@/features/messages/messagesSlice";
 import { setError } from "@/features/errors/errorsSlice";
-import { Button } from "@/components/ui/button";
-import {
-    Command,
-    CommandEmpty,
-    CommandInput,
-    CommandItem,
-    CommandList,
-} from "@/components/ui/command";
-import {
-    Form,
-    FormControl,
-    FormField,
-    FormItem,
-    FormLabel,
-    FormMessage,
-} from "@/components/ui/form";
 import {
     Select,
     SelectContent,
@@ -29,138 +8,18 @@ import {
     SelectTrigger,
     SelectValue,
 } from "@/components/ui/select";
-import {
-    Popover,
-    PopoverContent,
-    PopoverTrigger,
-} from "@/components/ui/popover";
-import { Calendar } from "./ui/calendar";
-import { Input } from "@/components/ui/input"
-import { CalendarIcon } from "lucide-react";
-import TimePicker from "./TimePicker";
-import { Textarea } from "./ui/textarea";
 import { useAppDispatch } from "@/app/hooks";
 import { useNavigate } from "react-router-dom";
-import { Checkbox } from "./ui/checkbox";
+import RequestNewDonationServiceForm from "./forms/RequestNewDonationServiceForm";
+import RequestNewMealForm from "./forms/RequestNewMealForm";
 
-const RequestTypeEnum = z.enum(["Donation", "Meal", "Service"]);
-type RequestTypeEnum = z.infer<typeof RequestTypeEnum>;
-const zipCodeRegex = /^[0-9]{5}(?:-[0-9]{4})?$/;
-
-const requestFormSchema = z.object({
-    request_type: RequestTypeEnum,
-    title: z.string()
-        .min(1, { message: "Enter a Title" })
-        .max(100, { message: "Title max length 100 characters" }),
-    recipient_id: z.number({ message: "Select a Recipient" }),
-    coordinator_id: z.number({ message: "Select a Coordinator" }),
-    notes: z.string()
-        .min(1, { message: "Enter Notes" })
-        .max(500, { message: "Notes max length 1000 characters" }),
-    allergies: z.string()
-        .max(100, { message: "Allergies max length 100 characters" })
-        .optional(),
-    date_range: z.object({ from: z.date(), to: z.date() }).optional(),
-    selected_days: z.array(z.number()).refine((value) => value.some((item) => item), {
-        message: "Select at least one day"
-    }).optional(),
-    // FIX: day, from, and to are seen as different Dates and need to combine into one
-    date_single: z.object({ day: z.date(), from: z.date(), to: z.date() }).optional(),
-    street_line: z.string()
-        .min(1, { message: "Enter a Street Address" })
-        .max(100, { message: "Street Address max length 100 characters" }),
-    city: z.string()
-        .min(1, { message: "Enter a City" })
-        .max(50, { message: "City max length 50 characters" }),
-    state: z.string()
-        .length(2, { message: "Use State abbreviation of 2 characters" })
-        .transform((val) => val.toUpperCase()),
-    zip_code: z.string()
-        .refine((val) => zipCodeRegex.test(val), { message: "Invalid ZIP code format (12345 / 12345-1234)" }),
-}).refine(data => {
-    if (data.date_range == undefined) {
-        return true
-    }
-    if (data.date_range?.from && data.date_range?.to > data.date_range?.from) {
-        return true;
-    } else {
-        return false;
-    }
-}, {
-    message: "Start Date must be before End Date",
-    path: ["date_range.from"],
-});
 
 function RequestNewForm() {
     const dispatch = useAppDispatch();
     const navigate = useNavigate();
 
     const [people, setPeople] = useState([]);
-    const [selectedRecipient, setSelectedRecipient] = useState(null);
-    const [selectedCoordinator, setSelectedCoordinator] = useState(null);
-    const [openRecipientSearch, setOpenRecipientSearch] = useState(false);
-    const [openCoordinatorSearch, setOpenCoordinatorSearch] = useState(false);
-    const [dateRangeFrom, setDateRangeFrom] = useState(false);
-    const [dateRangeTo, setDateRangeTo] = useState(false);
-
-    const today = new Date(new Date().setHours(0, 0, 0, 0));
-    const endDateRange = new Date(new Date().setMonth(today.getMonth() + 3));
-    const days = [
-        {
-            id: 1,
-            label: "Sunday",
-        },
-        {
-            id: 2,
-            label: "Monday",
-        },
-        {
-            id: 3,
-            label: "Tuesday",
-        },
-        {
-            id: 4,
-            label: "Wednesday",
-        },
-        {
-            id: 5,
-            label: "Thursday",
-        },
-        {
-            id: 6,
-            label: "Friday",
-        },
-        {
-            id: 7,
-            label: "Saturday",
-        },
-    ];
-    const form = useForm<z.infer<typeof requestFormSchema>>({
-        resolver: zodResolver(requestFormSchema),
-        defaultValues: {
-            request_type: "Meal",
-            title: "",
-            notes: "",
-            allergies: "",
-            date_range: {
-                from: today,
-                to: endDateRange,
-            },
-            selected_days: [],
-            date_single: {
-                day: today,
-                from: today,
-                to: today,
-            },
-            street_line: "",
-            city: "",
-            state: "",
-            zip_code: "",
-        },
-        shouldUnregister: true,
-    });
-    const { watch } = form;
-    const watchRequestTypeSelect = watch("request_type");
+    const [selectedRequestType, setSelectedRequestType] = useState("Meal");
 
     // get current list of people in organization to search through
     useEffect(() => {
@@ -207,367 +66,22 @@ function RequestNewForm() {
         }
     }, [])
 
-    function onSubmit(values: z.infer<typeof requestFormSchema>) {
-        console.log("SUBMITTED");
-        console.log(values);
-    }
-
     return <>
-        <Form {...form}>
-            <form onSubmit={form.handleSubmit(onSubmit)} className="flex flex-col gap-4 justify-center">
-                <FormField control={form.control} name="request_type" render={({ field }) => (
-                    <FormItem>
-                        <FormLabel>Type</FormLabel>
-                        <FormControl>
-                            <Select onValueChange={field.onChange} defaultValue={field.value}>
-                                <SelectTrigger className="w-[180px]">
-                                    <SelectValue placeholder="Select a Category" />
-                                </SelectTrigger>
-                                <SelectContent>
-                                    <SelectItem value="Donation">Donation</SelectItem>
-                                    <SelectItem value="Meal">Meal</SelectItem>
-                                    <SelectItem value="Service">Service</SelectItem>
-                                </SelectContent>
-                            </Select>
-                        </FormControl>
-                        <FormMessage />
-                    </FormItem>
-                )}
-                />
-                <FormField control={form.control} name="recipient_id" render={({ field }) => (
-                    <FormItem>
-                        <Popover open={openRecipientSearch} onOpenChange={setOpenRecipientSearch}>
-                            <PopoverTrigger asChild>
-                                <FormControl>
-                                    <Button variant={"outline"} role={"combobox"} >
-                                        {selectedRecipient ? selectedRecipient.name : "Select Recipient"}
-                                    </Button>
-                                </FormControl>
-                            </PopoverTrigger>
-                            <PopoverContent>
-                                <Command>
-                                    <CommandInput placeholder="Search for Recipient" />
-                                    <CommandList>
-                                        <CommandEmpty>No results found.</CommandEmpty>
-                                        {
-                                            people.map((person, index) => (
-                                                <CommandItem className="py-3" key={index} onSelect={() => {
-                                                    form.setValue("recipient_id", person.id);
-                                                    setSelectedRecipient(person)
-                                                    setOpenRecipientSearch(false)
-                                                }}>
-                                                    <>
-                                                        {person.name}
-                                                        <br />
-                                                        {person.email && person.email}
-                                                        {person.phone_number && <><br />{person.phone_number}</>}
-                                                    </>
-                                                </CommandItem>
-                                            ))}
-                                    </CommandList>
-                                </Command>
-                            </PopoverContent>
-                        </Popover>
-                        <FormMessage />
-                    </FormItem>
-                )}
-                />
-                <FormField control={form.control} name="coordinator_id" render={({ field }) => (
-                    <FormItem>
-                        <Popover open={openCoordinatorSearch} onOpenChange={setOpenCoordinatorSearch}>
-                            <PopoverTrigger asChild>
-                                <FormControl>
-                                    <Button variant={"outline"} role={"combobox"} >
-                                        {selectedCoordinator ? selectedCoordinator.name : "Select Coordinator"}
-                                    </Button>
-                                </FormControl>
-                            </PopoverTrigger>
-                            <PopoverContent>
-                                <Command>
-                                    <CommandInput placeholder="Search for Coordinator" />
-                                    <CommandList>
-                                        <CommandEmpty>No results found.</CommandEmpty>
-                                        {
-                                            people.map((person, index) => (
-                                                <CommandItem className="py-3" key={index} onSelect={() => {
-                                                    form.setValue("coordinator_id", person.id);
-                                                    setSelectedCoordinator(person)
-                                                    setOpenCoordinatorSearch(false)
-                                                }}>
-                                                    <>
-                                                        {person.name}
-                                                        <br />
-                                                        {person.email && person.email}
-                                                        {person.phone_number && <><br />{person.phone_number}</>}
-                                                    </>
-                                                </CommandItem>
-                                            ))}
-                                    </CommandList>
-                                </Command>
-                            </PopoverContent>
-                        </Popover>
-                        <FormMessage />
-                    </FormItem>
-                )}
-                />
-                <FormField control={form.control} name="title" render={({ field }) => (
-                    <FormItem>
-                        <FormLabel>Title</FormLabel>
-                        <FormControl>
-                            <Input maxLength={100} placeholder="" {...field} />
-                        </FormControl>
-                        <FormMessage />
-                    </FormItem>
-                )}
-                />
-                <FormField control={form.control} name="notes" render={({ field }) => (
-                    <FormItem>
-                        <FormLabel>Notes</FormLabel>
-                        <FormControl>
-                            <Textarea maxLength={500} placeholder="" {...field} />
-                        </FormControl>
-                        <FormMessage />
-                    </FormItem>
-                )}
-                />
-                {watchRequestTypeSelect == "Meal" && (
-                    <>
-                        <FormField control={form.control} name="allergies" render={({ field }) => (
-                            <FormItem>
-                                <FormLabel>Allergies</FormLabel>
-                                <FormControl>
-                                    <Input maxLength={100} placeholder="" {...field} />
-                                </FormControl>
-                                <FormMessage />
-                            </FormItem>
-                        )}
-                        />
-                        <FormField
-                            control={form.control} name="date_range.from" render={({ field }) => (
-                                <FormItem className="flex flex-col">
-                                    <FormLabel>Start Date</FormLabel>
-                                    {
-                                        // TODO: make popover close after date selection
-                                    }
-                                    <Popover open={dateRangeFrom} onOpenChange={setDateRangeFrom}>
-                                        <PopoverTrigger asChild>
-                                            <FormControl>
-                                                <Button variant={"outline"} className={cn(
-                                                    "w-[240px] justify-between pl-3 text-left font-normal",
-                                                    !field.value && "text-muted-foreground"
-                                                )}
-                                                >
-                                                    {field.value ? (
-                                                        format(field.value, "PP")
-                                                    ) : (
-                                                        <span>Pick a date</span>
-                                                    )}
-                                                    <CalendarIcon className="h-4 w-4" />
-                                                </Button>
-                                            </FormControl>
-                                        </PopoverTrigger>
-                                        <PopoverContent className="w-auto p-0" align="start">
-                                            <Calendar
-                                                mode="single"
-                                                selected={field.value}
-                                                onSelect={field.onChange}
-                                                disabled={(date) =>
-                                                    date < today || date > endDateRange
-                                                }
-                                            />
-                                        </PopoverContent>
-                                    </Popover>
-                                    <FormMessage />
-                                </FormItem>
-                            )}
-                        />
-                        <FormField
-                            control={form.control} name="date_range.to" render={({ field }) => (
-                                <FormItem className="flex flex-col">
-                                    <FormLabel>End Date</FormLabel>
-                                    <Popover open={dateRangeTo} onOpenChange={setDateRangeTo}>
-                                        <PopoverTrigger asChild>
-                                            <FormControl>
-                                                <Button variant={"outline"} className={cn(
-                                                    "w-[240px] justify-between pl-3 text-left font-normal",
-                                                    !field.value && "text-muted-foreground"
-                                                )}
-                                                >
-                                                    {field.value ? (
-                                                        format(field.value, "PP")
-                                                    ) : (
-                                                        <span>Pick a date</span>
-                                                    )}
-                                                    <CalendarIcon className="h-4 w-4" />
-                                                </Button>
-                                            </FormControl>
-                                        </PopoverTrigger>
-                                        <PopoverContent className="w-auto p-0" align="start">
-                                            <Calendar
-                                                mode="single"
-                                                selected={field.value}
-                                                onSelect={field.onChange}
-                                                disabled={(date) =>
-                                                    date < today || date > endDateRange
-                                                }
-                                            />
-                                        </PopoverContent>
-                                    </Popover>
-                                    <FormMessage />
-                                </FormItem>
-                            )}
-                        />
-                        <FormField
-                            control={form.control}
-                            name="selected_days"
-                            render={() => (
-                                <FormItem>
-                                    <FormLabel>Delivery Days</FormLabel>
-                                    {days.map((day) => (
-                                        <FormField
-                                            key={day.id}
-                                            control={form.control}
-                                            name="selected_days"
-                                            render={({ field }) => {
-                                                return (
-                                                    <FormItem
-                                                        key={day.id}
-                                                        className="flex flex-row items-start space-x-3 space-y-0"
-                                                    >
-                                                        <FormControl>
-                                                            <Checkbox
-                                                                checked={field.value?.includes(day.id)}
-                                                                onCheckedChange={(checked) => {
-                                                                    return checked
-                                                                        ? field.onChange([...field.value, day.id])
-                                                                        : field.onChange(
-                                                                            field.value?.filter(
-                                                                                (value) => value !== day.id
-                                                                            )
-                                                                        )
-                                                                }}
-                                                            />
-                                                        </FormControl>
-                                                        <FormLabel className="font-normal">
-                                                            {day.label}
-                                                        </FormLabel>
-                                                    </FormItem>
-                                                )
-                                            }}
-                                        />
-                                    ))}
-                                    <FormMessage />
-                                </FormItem>
-                            )}
-                        />
-                    </>
-                )}
-                {watchRequestTypeSelect != "Meal" && (
-                    <>
-                        <FormField control={form.control} name="date_single.day" render={({ field }) => (
-                            <FormItem className="flex flex-col">
-                                <FormLabel>Start Date</FormLabel>
-                                <Popover>
-                                    <PopoverTrigger asChild>
-                                        <FormControl>
-                                            <Button variant={"outline"} className={cn(
-                                                "w-[240px] justify-between pl-3 text-left font-normal",
-                                                !field.value && "text-muted-foreground"
-                                            )}
-                                            >
-                                                {field.value ? (
-                                                    format(field.value, "PP")
-                                                ) : (
-                                                    <span>Pick a date</span>
-                                                )}
-                                                <CalendarIcon className="h-4 w-4" />
-                                            </Button>
-                                        </FormControl>
-                                    </PopoverTrigger>
-                                    <PopoverContent className="w-auto p-0" align="start">
-                                        <Calendar
-                                            mode="single"
-                                            selected={field.value}
-                                            onSelect={field.onChange}
-                                            disabled={(date) =>
-                                                date < today || date > endDateRange
-                                            }
-                                        />
-                                    </PopoverContent>
-                                </Popover>
-                                <FormMessage />
-                            </FormItem>
-                        )}
-                        />
-                        <FormField control={form.control} name="date_single.from" render={({ field }) => (
-                            <FormItem className="flex flex-col">
-                                <FormLabel>Start Time</FormLabel>
-                                <FormControl>
-                                    <div className="p-3">
-                                        <TimePicker date={field.value} setDate={field.onChange} />
-                                    </div>
-                                </FormControl>
-                                <FormMessage />
-                            </FormItem>
-                        )}
-                        />
-                        <FormField control={form.control} name="date_single.to" render={({ field }) => (
-                            <FormItem className="flex flex-col">
-                                <FormLabel>End Time</FormLabel>
-                                <FormControl>
-                                    <div className="p-3">
-                                        <TimePicker date={field.value} setDate={field.onChange} />
-                                    </div>
-                                </FormControl>
-                                <FormMessage />
-                            </FormItem>
-                        )}
-                        />
-                    </>
-                )}
-                <FormField control={form.control} name="street_line" render={({ field }) => (
-                    <FormItem>
-                        <FormLabel>Street Address</FormLabel>
-                        <FormControl>
-                            <Input maxLength={100} placeholder="" {...field} />
-                        </FormControl>
-                        <FormMessage />
-                    </FormItem>
-                )}
-                />
-                <FormField control={form.control} name="city" render={({ field }) => (
-                    <FormItem>
-                        <FormLabel>City</FormLabel>
-                        <FormControl>
-                            <Input maxLength={50} placeholder="" {...field} />
-                        </FormControl>
-                        <FormMessage />
-                    </FormItem>
-                )}
-                />
-                <FormField control={form.control} name="state" render={({ field }) => (
-                    <FormItem>
-                        <FormLabel>State</FormLabel>
-                        <FormControl>
-                            <Input maxLength={2} placeholder="" {...field} />
-                        </FormControl>
-                        <FormMessage />
-                    </FormItem>
-                )}
-                />
-                <FormField control={form.control} name="zip_code" render={({ field }) => (
-                    <FormItem>
-                        <FormLabel>ZIP Code</FormLabel>
-                        <FormControl>
-                            <Input maxLength={10} placeholder="" {...field} />
-                        </FormControl>
-                        <FormMessage />
-                    </FormItem>
-                )}
-                />
-                <Button type="submit">Submit</Button>
-            </form>
-        </Form >
+        <div className="mb-5">
+            <Select onValueChange={setSelectedRequestType} defaultValue={selectedRequestType}>
+                <SelectTrigger className="w-[180px]">
+                    <SelectValue placeholder="Select a Category" />
+                </SelectTrigger>
+                <SelectContent>
+                    <SelectItem value="Donation">Donation</SelectItem>
+                    <SelectItem value="Meal">Meal</SelectItem>
+                    <SelectItem value="Service">Service</SelectItem>
+                </SelectContent>
+            </Select>
+        </div>
+
+        {(selectedRequestType == "Donation" || selectedRequestType == "Service") && <RequestNewDonationServiceForm requestType={selectedRequestType} people={people} />}
+        {selectedRequestType == "Meal" && <RequestNewMealForm requestType={selectedRequestType} people={people} />}
     </>
 }
